@abhijain1204fujitsu 2025-08-08
> Hi @nikhilfujitsu, these are great perf gains!
> 
> Could you clarify on which CPU(s) did you witness these perf improvements? And do you have steps to reproduce them?

Hi @Rohanjames1997 the results were captured using AWS Graviton 3e Machine (c7gn), you can try to run on the same