@Rohanjames1997 2025-08-08
Thanks @abhijain1204fujitsu , I will try it out later. 
If you could share the steps to repro the speedup - like `onnxruntime_perf_test` or the like, that would help too.