@nikhilfujitsu 2025-09-17
> > @edgchen1 sir please support this PR further. Thanks.
> 
> The ONNX submodule update doesn't seem to be removed.

Updated the cmake/externel/onnx.