@hariharans29 2025-09-30
Please run onnxruntime_mlas_test in your env with your change and report any failures if any here.