@nikhilfujitsu 2025-10-08
> > Build and install the wheel of Onnxruntime. Download the INT/Float E5 model from Hugging Face. Convert the model to FP16 ONNX. Create a Python Script to Run Inference. To Analyze Operator-Level Timings use Session option.
> > We used AWS Graviton 3e Machine (c7gn) with 64 cores.
> 
> thanks. would you mind also updating the PR description to include this information, so it is easier to find later?
> 
> also, regarding these measurements:
> 
> > 2.Enable fp16 to fp32 cast(performance improved from 15sec -> 0.27sec)
> > 3.Added eigen::fp16 support in layer normalization (performance improved from 22sec to 0.6sec)
> > 4.Enable fp16 Transpose call in mlas(performance improved from 3sec -> 0.47sec​)
> 
> could you please clarify what is being measured? are the timings for different sections of the FP16 E5 model?

I have updated the Description.
During the inference operation only the time is captured. Yes, the timings are for the different ops for the E5 model.