@nikhilfujitsu 2025-10-16
Hi @hariharans29 sir,
Seems like build is okay but issues with CI.
The repository's CI requirements file pins onnxscript==0.3.2 (tools/ci_build/github/linux/python/requirements.txt in ref 4f238652eb39bc8381fa41b4744aa647961ad285), which appears incompatible with the runner's torch version.